+++ R/utils-extract-media.R
@@ -71,3 +71,36 @@ fix_ids <- function(ids) {
                    .else = as.character)
 
 }
+
+#' Quick datetime conversion
+#'
+#' Searches for datetime variables and converts them to `POSIXct` via \pkg{lubridate}.
+#' @param response The input object. Must be `data.frame`(ish) or named `list`.
+#' @return The same object with converted datetimes
+#' @importFrom lubridate ymd_hms
+#' @importFrom dplyr mutate_at
+#' @importFrom purrr map_at
+#' @keywords internal
+fix_datetime <- function(response) {
+  if (!inherits(response, c("data.frame", "list"))) {
+    stop("Object type not supported, must inherit from data.frame or list")
+  }
+  datevars <- c(
+    "first_aired", "updated_at", "listed_at", "last_watched_at", "last_updated_at",
+    "last_collected_at", "rated_at", "friends_at", "followed_at", "collected_at",
+    "joined_at", "watched_at"
+  )
+
+  datevars <- datevars[datevars %in% names(response)]
+
+  if (tibble::has_name(response, "released")) {
+    response$released <- lubridate::as_datetime(response$released)
+  }
+
+  if (inherits(response, "data.frame")) {
+    response %>%
+      dplyr::mutate_at(.vars = dplyr::vars(datevars), lubridate::ymd_hms)
+  } else {
+    purrr::map_at(response, datevars, lubridate::ymd_hms)
+  }
+}